+++ src/search/mod.rs
@@ -1,4 +1,5 @@
 pub mod ddg;
+pub mod geeksforgeeks;
 pub mod stackexchange;
 pub mod stackoverflow;
 pub mod utils;